Boot-mon-server is refactored, docker part is added, mvn wrapper.
