+++ boot-mon-server/src/main/docker/Dockerfile
@@ -1,4 +1,4 @@
-FROM java:8
+FROM java:8-jre-alpine
 VOLUME /tmp
 ADD boot-mon-server-1-SNAPSHOT.jar boot-mon-server.jar
 RUN bash -c 'touch /boot-mon-server.jar'
