+++ boot-mon-server/src/main/java/com/iyzico/bootmon/server/repository/BootmonClientRepository.java
@@ -1,7 +1,8 @@
-package com.iyzico.bootmon.server;
+package com.iyzico.bootmon.server.repository;
 
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.iyzico.bootmon.server.dto.shared.BootmonClient;
 
 import java.io.IOException;
 import java.util.Optional;
