+++ boot-mon-server/src/main/java/com/iyzico/bootmon/server/service/core/impl/BootmonClientServiceImpl.java
@@ -1,6 +1,9 @@
-package com.iyzico.bootmon.server;
+package com.iyzico.bootmon.server.service.core.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.iyzico.bootmon.server.dto.shared.BootmonClient;
+import com.iyzico.bootmon.server.repository.BootmonClientRepository;
+import com.iyzico.bootmon.server.service.core.BootmonClientService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
